operator: remove pod list of an entire cluster
Doing a pod list every 30 minutes can be brutal for kube-apiserver,
specially if it is not running with a limit of the number of pods that
should be returned. Having a watcher that receives notifications for the
pods running as well as the CEPs, is enough to provide a GC for
Cilium Identities and CEPs.

Signed-off-by: André Martins <andre@cilium.io>