pkg/k8s: cast to networkpolicy even if obj is cache.DeletedFinalStateUnknown
When casting to a networkpolicy, the k8s watcher might send a
cache.DeletedFinalStateUnknown object to the event handler. Since the
converter is not expecting this type, it will cause a warning to be
printed in the logs which will be unexpected for users. This commit
adds the ability for the function converter to cast the objects from
cache.DeletedFinalStateUnknown to avoid printing such warning.

Signed-off-by: André Martins <andre@cilium.io>