bpf: Check native-routing-cidr in BPF masquerade
This commit extends the BPF masquerade by adding a check for whether
a dst IPv4 addr is in native-routing-cidr.

The check fixes a problem when a node hasn't received an ipcache update
for a pod running on a remote node, and thus a packet sent from a local
pod to that pod is SNAT'd. This happens because the dst is considered
as WORLD_ID until the update has been received.

Signed-off-by: Martynas Pumputis <m@lambda.lt>